Rename common-substrate to common-algorand, update changelogs, update CI
