+++ .github/actions/create-release/gh-release-script.js
@@ -68,7 +68,7 @@ async function publishRelease(releaseInfo) {
         },
         owner: 'subquery',
         name: `[${version}] ${repoName}`,
-        repo: 'subql',
+        repo: 'subql-algorand',
         tag_name: `${packageName[1]}/${version}`,
         body: releaseInfo
     }).catch( err => {
