+++ .github/workflows/gh-release.yml
@@ -22,17 +22,7 @@ jobs:
       #Create github releases
       - uses: ./.github/actions/create-release
         with:
-          package-path: packages/cli
-          repo-token: ${{ secrets.REPO_TOKEN }}
-
-      - uses: ./.github/actions/create-release
-        with:
-          package-path: packages/common
-          repo-token: ${{ secrets.REPO_TOKEN }}
-
-      - uses: ./.github/actions/create-release
-        with:
-          package-path: packages/common-substrate
+          package-path: packages/common-algorand
           repo-token: ${{ secrets.REPO_TOKEN }}
 
       - uses: ./.github/actions/create-release
@@ -42,20 +32,5 @@ jobs:
 
       - uses: ./.github/actions/create-release
         with:
-          package-path: packages/query
-          repo-token: ${{ secrets.REPO_TOKEN }}
-
-      - uses: ./.github/actions/create-release
-        with:
           package-path: packages/types
-          repo-token: ${{ secrets.REPO_TOKEN }}
-
-      - uses: ./.github/actions/create-release
-        with:
-          package-path: packages/utils
-          repo-token: ${{ secrets.REPO_TOKEN }}
-
-      - uses: ./.github/actions/create-release
-        with:
-          package-path: packages/validator
           repo-token: ${{ secrets.REPO_TOKEN }}
