--- .github/workflows/node-avalanche-docker.yml
@@ -1,60 +1,0 @@
-name: "Node-Avalanche-to-docker"
-on:
-  workflow_dispatch:
-    inputs:
-      isLatest:
-        description: 'Add latest tag'
-        default: 'true'
-        required: true
-
-jobs:
-  node-build-push-docker:
-
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          fetch-depth: 100
-          token: ${{ secrets.REPO_TOKEN }}
-
-      - name: Set up QEMU
-        uses: docker/setup-qemu-action@v1
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v1
-
-      - name: Login to DockerHub
-        uses: docker/login-action@v1
-        with:
-          username: onfinality
-          password: ${{ secrets.DOCKERHUB_TOKEN }}
-
-      ## node
-      - name: Get updated node version
-        id: get-node-version
-        run: | #TODO
-          sh .github/workflows/scripts/nodeAvalancheVersion.sh
-
-      - name: Build and push
-        if: github.event.inputs.isLatest == 'false'
-        uses: docker/build-push-action@v2
-        with:
-          push: true
-          platforms: arm64,amd64
-          file: ./packages/node-avalanche/Dockerfile
-          tags: onfinality/subql-node-avalanche:v${{ steps.get-node-version.outputs.NODE_VERSION }}
-          build-args: RELEASE_VERSION=${{ steps.get-node-version.outputs.NODE_VERSION }}
-
-      - name: Build and push
-        if: github.event.inputs.isLatest == 'true'
-        uses: docker/build-push-action@v2
-        with:
-          push: true
-          platforms: arm64,amd64
-          file: ./packages/node-avalanche/Dockerfile
-          tags: onfinality/subql-node-avalanche:v${{ steps.get-node-version.outputs.NODE_VERSION }},onfinality/subql-node-avalanche:latest
-          build-args: RELEASE_VERSION=${{ steps.get-node-version.outputs.NODE_VERSION }}
-
-      - name: Image digest
-        run: echo ${{ steps.docker_build.outputs.digest }}
-
