+++ .github/workflows/prerelease.yml
@@ -22,7 +22,7 @@ jobs:
   Bump-Prerelease-Publish:
     name: Bump-Prerelease-Publish
     needs: pre-ci
-    if: "!startsWith(github.event.head_commit.message, '[SKIP CI]') && !startsWith(github.event.head_commit.message, '[release]') && github.repository == 'subquery/subql'"
+    if: "!startsWith(github.event.head_commit.message, '[SKIP CI]') && !startsWith(github.event.head_commit.message, '[release]') && github.repository == 'subquery/subql-algorand'"
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
@@ -41,39 +41,14 @@ jobs:
           paths: packages/types
 
       - uses: marceloprado/has-changed-path@v1
-        id: changed-common
+        id: changed-common-algorand
         with:
-          paths: packages/common
-
-      - uses: marceloprado/has-changed-path@v1
-        id: changed-common-substrate
-        with:
-          paths: packages/common-substrate
-
-      - uses: marceloprado/has-changed-path@v1
-        id: changed-cli
-        with:
-          paths: packages/cli
+          paths: packages/common-algorand
 
       - uses: marceloprado/has-changed-path@v1
         id: changed-node
         with:
           paths: packages/node
-
-      - uses: marceloprado/has-changed-path@v1
-        id: changed-query
-        with:
-          paths: packages/query
-
-      - uses: marceloprado/has-changed-path@v1
-        id: changed-utils
-        with:
-          paths: packages/utils
-
-      - uses: marceloprado/has-changed-path@v1
-        id: changed-validator
-        with:
-          paths: packages/validator
 
       - run: yarn
 
@@ -81,12 +56,6 @@ jobs:
         run: yarn workspaces foreach run build
 
       #Add remove-stable-version
-      - name: Bump utils & deploy
-        if: steps.changed-utils.outputs.changed == 'true'
-        uses: ./.github/actions/create-prerelease
-        with:
-          package-path: packages/utils
-          npm-token: ${{ secrets.NPM_TOKEN }}
 
       - name: Bump types & deploy
         if: steps.changed-types.outputs.changed == 'true'
@@ -95,18 +64,11 @@ jobs:
           package-path: packages/types
           npm-token: ${{ secrets.NPM_TOKEN }}
 
-      - name: Bump common & deploy
-        if: steps.changed-common.outputs.changed == 'true'
+      - name: Bump common algorand & deploy
+        if: steps.changed-common-algorand.outputs.changed == 'true'
         uses: ./.github/actions/create-prerelease
         with:
-          package-path: packages/common
-          npm-token: ${{ secrets.NPM_TOKEN }}
-
-      - name: Bump common substrate & deploy
-        if: steps.changed-common-substrate.outputs.changed == 'true'
-        uses: ./.github/actions/create-prerelease
-        with:
-          package-path: packages/common-substrate
+          package-path: packages/common-algorand
           npm-token: ${{ secrets.NPM_TOKEN }}
 
       - name: Bump node & deploy
@@ -114,27 +76,6 @@ jobs:
         uses: ./.github/actions/create-prerelease
         with:
           package-path: packages/node
-          npm-token: ${{ secrets.NPM_TOKEN }}
-
-      - name: Bump query & deploy
-        if: steps.changed-query.outputs.changed == 'true'
-        uses: ./.github/actions/create-prerelease
-        with:
-          package-path: packages/query
-          npm-token: ${{ secrets.NPM_TOKEN }}
-
-      - name: Bump validator & deploy
-        if: steps.changed-validator.outputs.changed == 'true'
-        uses: ./.github/actions/create-prerelease
-        with:
-          package-path: packages/validator
-          npm-token: ${{ secrets.NPM_TOKEN }}
-
-      - name: Bump cli & deploy
-        if: steps.changed-cli.outputs.changed == 'true'
-        uses: ./.github/actions/create-prerelease
-        with:
-          package-path: packages/cli
           npm-token: ${{ secrets.NPM_TOKEN }}
 
       - name: Commit changes
