--- .github/workflows/query-docker.yml
@@ -1,60 +1,0 @@
-name: "Query-to-docker"
-on:
-  workflow_dispatch:
-    inputs:
-      isLatest:
-        description: 'Add latest tag'
-        default: 'true'
-        require: true
-
-jobs:
-  query-build-push-docker:
-
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          fetch-depth: 100
-          token: ${{ secrets.REPO_TOKEN }}
-
-      - name: Set up QEMU
-        uses: docker/setup-qemu-action@v1
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v1
-
-      - name: Login to DockerHub
-        uses: docker/login-action@v1
-        with:
-          username: onfinality
-          password: ${{ secrets.DOCKERHUB_TOKEN }}
-
-      ## query
-      - name: Get updated query version
-        id: get-query-version
-        run: |
-          sh .github/workflows/scripts/queryVersion.sh
-
-      - name: Build and push
-        if: github.event.inputs.isLatest == 'false'
-        uses: docker/build-push-action@v2
-        with:
-          push: true
-          platforms: arm64,amd64
-          file: ./packages/query/Dockerfile
-          tags: onfinality/subql-query:v${{ steps.get-query-version.outputs.QUERY_VERSION }}
-          build-args: RELEASE_VERSION=${{ steps.get-query-version.outputs.QUERY_VERSION }}
-
-      - name: Build and push
-        if: github.event.inputs.isLatest == 'true'
-        uses: docker/build-push-action@v2
-        with:
-          push: true
-          platforms: arm64,amd64
-          file: ./packages/query/Dockerfile
-          tags: onfinality/subql-query:v${{ steps.get-query-version.outputs.QUERY_VERSION }},onfinality/subql-query:latest
-          build-args: RELEASE_VERSION=${{ steps.get-query-version.outputs.QUERY_VERSION }}
-
-      - name: Image digest
-        if: steps.changed-query.outputs.changed == 'true'
-        run: echo ${{ steps.docker_build.outputs.digest }}
