+++ .github/workflows/scripts/gh-release-script.js
@@ -62,7 +62,7 @@ async function publishRelease(releaseInfo) {
         },
         owner: 'subql',
         name: `[${version}] ${repoName}`,
-        repo: 'subql',
+        repo: 'subql-algorand',
         tag_name: `${repoTagName[1]}/${version}`,
         body: releaseInfo
     }).catch( err => {
