+++ packages/common-algorand/CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+All logs must start with the format: [x.y.z] - yyyy-mm-dd
+
 ## [Unreleased]
 
-Initial Algorand support.
+Initial Algorand support
 
-[Unreleased]: https://github.com/subquery/subql/compare/node/HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/common-algorand/HEAD
