+++ packages/common-algorand/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@subql/common-substrate",
-  "version": "1.1.0",
+  "name": "@subql/common-algorand",
+  "version": "0.0.1",
   "description": "",
   "scripts": {
     "build": "rm -rf dist && tsc -b",
