+++ packages/common-algorand/src/project/index.ts
