+++ packages/common-algorand/src/project/load.ts
