+++ packages/common-algorand/src/project/types.ts
