+++ packages/common-algorand/src/project/utils.ts
