+++ packages/common-algorand/src/project/validation/is-string-or-object.validation.ts
