+++ packages/common-algorand/src/project/versioned/ProjectManifestVersioned.ts
