+++ packages/common-algorand/test/project.yaml
