+++ packages/common-algorand/test/project_falsy.yaml
