+++ packages/common-algorand/test/project_falsy_array.yaml
