+++ packages/common-algorand/test/project_invalid_version.yaml
