+++ packages/common-algorand/tsconfig.json
