--- packages/common-substrate/CHANGELOG.md
@@ -1,45 +1,0 @@
-# Changelog
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-All logs must start with the format: [x.y.z] - yyyy-mm-dd
-
-## [Unreleased]
-
-## [1.1.0] - 2022-05-31
-### Changed
-- Remove `contract-processors` and types improvements (#1012)
-
-## [1.0.0] - 2022-05-11
-- Major release
-
-## [0.5.0] - 2022-05-11
-### Added
-- Add test for project with ds processor assets (#1000)
-- Add method to check templates is Substrate ds (#1001)
-
-## [0.4.0] - 2022-05-02
-### Added
-- Add utils package (#928)
-
-## [0.3.1] - 2022-04-27
-### Changed
-- Add missing vm2 dependency (#919)
-
-## [0.3.0] - 2022-04-06
-### Added
-- Add manifest 1.0.0 (#845)
-### Fixed
-- Fix validation for genesisHash or chainId with empty string (#883)
-
-## [0.2.0] - 2022-04-04
-### Changed
-- Update to use `vm2`(#869)
-
-## [0.1.0] - 2022-03-01
-### Added
-- init commit
-
-[Unreleased]: https://github.com/subquery/subql/compare/common-substrate/0.1.0...HEAD
