+++ packages/node/Dockerfile
@@ -2,7 +2,7 @@
 FROM node:16 as builder
 ARG RELEASE_VERSION
 ENTRYPOINT ["subql-node"]
-RUN npm i -g --unsafe-perm @subql/node@${RELEASE_VERSION}
+RUN npm i -g --unsafe-perm @subql/node-algorand${RELEASE_VERSION}
 
 FROM node:16-alpine
 ENV TZ utc
@@ -10,5 +10,5 @@ ENV TZ utc
 RUN apk add --no-cache tini git curl
 COPY --from=builder /usr/local/lib/node_modules /usr/local/lib/node_modules
 
-ENTRYPOINT ["/sbin/tini", "--", "/usr/local/lib/node_modules/@subql/node/bin/run"]
+ENTRYPOINT ["/sbin/tini", "--", "/usr/local/lib/node_modules/@subql/node-algorand/bin/run"]
 CMD ["-f","/app"]
