+++ packages/node/src/configure/configure.module.ts
@@ -5,7 +5,7 @@ import assert from 'assert';
 import path from 'path';
 import { DynamicModule, Global, Module } from '@nestjs/common';
 import { getProjectRootAndManifest, IPFS_REGEX } from '@subql/common';
-import { AlgorandProjectNetworkConfig } from '@subql/common-substrate';
+import { AlgorandProjectNetworkConfig } from '@subql/common-algorand';
 import { camelCase, last, omitBy, isNil } from 'lodash';
 import { getLogger, setLevel } from '../utils/logger';
 import { getYargsOption } from '../yargs';
